# song-selection-backend/src/main/openapi/song-api.yml
openapi: 3.0.0
info:
  title: Liedjes Selectie API
  description: API voor het beheren van christelijke liederen en liedkeuzes.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Development server
...
paths:
  /songs/search:
    get:
      summary: Zoek liedjes op basis van een zoekterm
      parameters:
        - in: query
          name: q
          schema:
            type: string
          required: true
          description: De zoekterm om liedjes op te halen
      responses:
        '200':
          description: Een lijst van liedjes die overeenkomen met de zoekterm
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Song'
        '400':
          description: Ongeldige zoekterm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Song:
      type: object
      properties:
        number:
          type: integer
        title:
          type: string
        bundle:
          type: string
    Error:
      type: object
      properties:
        error:
          type: string